fix(patina): keep no-deprecated-filter under source-length guard
Split the rule's unit tests into a sibling no_deprecated_filter_tests.rs
file (matching the script_tests.rs precedent) so the rule module drops
from 406 to 260 lines, and move the new translation entries into
i18n_supplemental_extra.rs to keep i18n_supplemental.rs at its 347-line
baseline. Both files now sit under the 350-line guard the
source-file-lengths tooling test enforces against the base branch.
